refactoring Dot objects in spec examples
